move not-mocks out of mock.rs - part 2, reorganize mock, generic Network
