+++ consensus/src/testing/mock/crypto.rs
@@ -1,0 +1,74 @@
+use crate::{
+    Index, KeyBox as KeyBoxT, MultiKeychain as MultiKeychainT, NodeCount, NodeIndex,
+    PartialMultisignature as PartialMultisignatureT,
+};
+use async_trait::async_trait;
+use codec::{Decode, Encode};
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Encode, Decode)]
+pub struct Signature {}
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Encode, Decode)]
+pub struct PartialMultisignature {
+    signed_by: Vec<NodeIndex>,
+}
+
+impl PartialMultisignatureT for PartialMultisignature {
+    type Signature = Signature;
+    fn add_signature(self, _: &Self::Signature, index: NodeIndex) -> Self {
+        let Self { mut signed_by } = self;
+        for id in &signed_by {
+            if *id == index {
+                return Self { signed_by };
+            }
+        }
+        signed_by.push(index);
+        Self { signed_by }
+    }
+}
+
+#[derive(Clone)]
+pub struct KeyBox {
+    count: NodeCount,
+    ix: NodeIndex,
+}
+
+impl KeyBox {
+    pub fn new(count: NodeCount, ix: NodeIndex) -> Self {
+        KeyBox { count, ix }
+    }
+}
+
+impl Index for KeyBox {
+    fn index(&self) -> NodeIndex {
+        self.ix
+    }
+}
+
+#[async_trait]
+impl KeyBoxT for KeyBox {
+    type Signature = Signature;
+
+    fn node_count(&self) -> NodeCount {
+        self.count
+    }
+
+    async fn sign(&self, _msg: &[u8]) -> Signature {
+        Signature {}
+    }
+
+    fn verify(&self, _msg: &[u8], _sgn: &Signature, _index: NodeIndex) -> bool {
+        true
+    }
+}
+
+impl MultiKeychainT for KeyBox {
+    type PartialMultisignature = PartialMultisignature;
+    fn from_signature(&self, _: &Self::Signature, index: NodeIndex) -> Self::PartialMultisignature {
+        let signed_by = vec![index];
+        PartialMultisignature { signed_by }
+    }
+    fn is_complete(&self, _: &[u8], partial: &Self::PartialMultisignature) -> bool {
+        (self.count * 2) / 3 < NodeCount(partial.signed_by.len())
+    }
+}
